All script types registered in registerQDaqTypes().
