+++ designer/qdaqconsoletabplugin.cpp
@@ -1,0 +1,6 @@
+#include "qdaqconsoletabplugin.h"
+
+QDaqConsoleTabPlugin::QDaqConsoleTabPlugin()
+{
+
+}
