+++ designer/qdaqconsoletabplugin.h
@@ -1,0 +1,11 @@
+#ifndef QDAQCONSOLETABPLUGIN_H
+#define QDAQCONSOLETABPLUGIN_H
+
+
+class QDaqConsoleTabPlugin
+{
+public:
+    QDaqConsoleTabPlugin();
+};
+
+#endif // QDAQCONSOLETABPLUGIN_H
